Fixed bugs, still need to fix post reviews

+++ app/models/photocard.py
@@ -9,7 +9,7 @@ class Photocard(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     user_id = db.Column(db.Integer, db.ForeignKey(add_prefix_for_prod('users.id')))
     listing_name = db.Column(db.String, nullable=False)
-    price = db.Column(db.Numeric(10,2), nullable=False)
+    price = db.Column(db.Integer, nullable=False)
     description = db.Column(db.Text, nullable=False)
     photocard_image = db.Column(db.String, nullable=False)
